+++ CHANGELOG.md
@@ -40,6 +40,8 @@ and this project adheres to
   [$3634](https://github.com/openfn/lightning/issues/3634)
 - View and switch versions
   [#3819](https://github.com/openfn/lightning/issues/3819)
+- Run viewer panel in collaborative IDE with Run, Log, Input, and Output tabs
+  [#3844](https://github.com/OpenFn/lightning/issues/3844)
 
 ### Changed
 
@@ -53,6 +55,10 @@ and this project adheres to
 
 ### Fixed
 
+- Prevent manual run panel from opening via keystroke if running isn't possible
+  (permissions or snapshot)
+- Default Next Input for Cron-Triggered Workflows
+  [#3856](https://github.com/OpenFn/lightning/issues/3856)
 - Error toasts not appearing when workflow save fails (validation errors,
   permission denied, etc.) in collaborative editor
   [#3843](https://github.com/OpenFn/lightning/issues/3843)